execに対してファイル分割を行った
