+++ src/ast/exec_ast_cmd.c
@@ -6,15 +6,11 @@
 /*   By: tishihar <tishihar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 16:07:58 by tishihar          #+#    #+#             */
-/*   Updated: 2025/04/27 19:10:47 by tishihar         ###   ########.fr       */
+/*   Updated: 2025/04/27 19:52:28 by tishihar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-static void	setup_child_fd(int fd_in, int fd_out);
-static void	check_fd(int fd_in, int fd_out);
-static int	validate_cmd_path(t_ast *node);
 
 // exec_ast_cmd() is used as follows.
 // step1. if there is a redirect arrays, replace fd_in and fd_out.
@@ -37,14 +33,7 @@ void	exec_right_cmd(t_ast *node, int fd_in, t_pids *pids)
 		exit_f("fork failed.");
 	else if (pid == 0)
 	{
-		set_exec_child_handler();
-		check_fd(fd_in, fd_out);
-		setup_child_fd(fd_in, fd_out);
-		if (validate_cmd_path(node))
-			exit(*node->u_data.cmd.stp);
-		execve(node->u_data.cmd.path, node->u_data.cmd.argv,
-			node->u_data.cmd.envp);
-		exit_f("execve failed");
+		exec_cmd(node, fd_in, fd_out);
 	}
 	else
 	{
@@ -73,15 +62,8 @@ void	exec_left_cmd(t_ast *node, int fd_in, int fd_pipe[], t_pids *pids)
 		exit_f("fork failed.");
 	else if (pid == 0)
 	{
-		set_exec_child_handler();
 		close(fd_pipe[0]);
-		check_fd(fd_in, fd_out);
-		setup_child_fd(fd_in, fd_out);
-		if (validate_cmd_path(node))
-			exit(*node->u_data.cmd.stp);
-		execve(node->u_data.cmd.path, node->u_data.cmd.argv,
-			node->u_data.cmd.envp);
-		exit_f("execve failed");
+		exec_cmd(node, fd_in, fd_out);
 	}
 	else
 	{
@@ -90,60 +72,4 @@ void	exec_left_cmd(t_ast *node, int fd_in, int fd_pipe[], t_pids *pids)
 		close(fd_pipe[1]);
 		pids_push_back(pids, pid);
 	}
-}
-
-// this func() determines if the input is correct.
-static void	check_fd(int fd_in, int fd_out)
-{
-	if (fd_in == -1)
-	{
-		perror("fd_in is not open");
-		exit(EXIT_FAILURE);
-	}
-	if (fd_out == -1)
-	{
-		perror("fd_out is not open");
-		exit(EXIT_FAILURE);
-	}
-}
-
-// this func() is set "fd_in" to STDIN, and "fd_out" to STD_OUT
-// if in or out isn't right shape, this is not the case.
-// The case where fd_in is -1 here is the case where the redirection is -1.
-// THE TRUE CASE: fd_out is appropriate, and, fd_out is not STDOUT
-static void	setup_child_fd(int fd_in, int fd_out)
-{
-	if (fd_in != -1 && fd_in != STDIN_FILENO)
-	{
-		dup2(fd_in, STDIN_FILENO);
-		close(fd_in);
-	}
-	if (fd_out != -1 && fd_out != STDOUT_FILENO)
-	{
-		dup2(fd_out, STDOUT_FILENO);
-		close(fd_out);
-	}
-}
-
-static	int	validate_cmd_path(t_ast *node)
-{
-	struct stat	st;
-	char		*path;
-
-	path = node->u_data.cmd.path;
-	if (!path || access(path, F_OK) != 0)
-	{
-		ft_putstr_fd("Command not found.\n", STDERR_FILENO);
-		return (*(node->u_data.cmd.stp) = 127, 1);
-	}
-	if (access(path, X_OK) != 0)
-		return (perror(path), *(node->u_data.cmd.stp) = 126, 1);
-	if (stat(path, &st) != 0)
-		return (perror(path), *(node->u_data.cmd.stp) = 126, 1);
-	if (!S_ISREG(st.st_mode) && !S_ISLNK(st.st_mode))
-	{
-		ft_putstr_fd("Permission denied.\n", STDERR_FILENO);
-		return (*(node->u_data.cmd.stp) = 126, 1);
-	}
-	return (0);
 }
